Convert cas list into a splay tree
The GKMZ algorithm always needs to construct a new internal CASN descriptor with
new internal CAS operations, because the CASN descriptor and the CAS operations
are consumed (installed to the refs).  This means that ordering the CASes in the
internal descriptor via the splay tree representation is fairly cheap.  When the
user gives the (external) CAS operations in either ascending or descending
order, the internal splay tree construction is linear time (`O(n)`).  Otherwise
the splay tree construction is linearithmic time (`O(n log n)`).

For the splay tree implementation, see e.g.

> Purely Functional Data Structures
> Chris Okasaki

and the `partition` function in the implementation of heaps using splay trees.