Allocate location ids in descending order
This should typically lead to slightly improved performance as it is typical
that locations closer to the root of a data structure are allocated before
locations further from the root.  So, when a data structure is traversed
starting from the root, the resuling splay tree should be right leaning and
slightly faster to construct and traverse from left to right.